Revert "Fix #3421: Separate MarkStoriesCompletedActivityTest into activity and fragment test files (#3470)"
This reverts commit 61d035fa6f2039cc9e380cf867283972fc91cdea.
