+++ app/src/main/AndroidManifest.xml
@@ -212,9 +212,6 @@
       android:label="@string/mark_stories_completed_activity_title"
       android:theme="@style/OppiaThemeWithoutActionBar" />
     <activity
-      android:name=".app.devoptions.markstoriescompleted.testing.MarkStoriesCompletedTestActivity"
-      android:theme="@style/OppiaThemeWithoutActionBar" />
-    <activity
       android:name=".app.devoptions.marktopicscompleted.MarkTopicsCompletedActivity"
       android:label="@string/mark_topics_completed_activity_title"
       android:theme="@style/OppiaThemeWithoutActionBar" />
