+++ app/src/main/res/layout/mark_stories_completed_fragment.xml
@@ -14,7 +14,6 @@
   </data>
 
   <androidx.constraintlayout.widget.ConstraintLayout
-    android:id="@+id/mark_stories_completed_fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white">
